+++ frontend/src/components/index.js
@@ -5,7 +5,7 @@ import { LocalNav } from './LocalNav';
 import { Spinner } from './Spinner';
 import { Thumbnail } from './Thumbnail';
 import { FoodTruck } from './FoodTruck';
-import { Modal } from './Modal';
+import { Modal } from './CustomModal';
 import { DetailFoodList } from './DetailFoodList';
 import { DetailInfo } from './DetailInfo';
 import { DetailReview } from './DetailReview';
